FuzzyAI/pyproject.toml
[tool.poetry]
name = "fuzzy"
version = "0.1.0"
description = "FuzzyAI is a Python library for fuzzying LLMs."
authors = ["CyberArk <cyber@cyberark.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.158"
motor = "^3.6"
pydantic = "^2.5.3"
aiofiles = "^23.2.1"
aiohttp = {version = "^3.9.1", extras = ["speedups"]}
transformers = "^4.37.2"
sentence-transformers = "^2.6.1"
fschat = "^0.2.36"
pygad = "^3.3.1"
backoff = "^2.2.1"
art = "^6.1"
pandas = "^2.2.2"
openai = "^1.34.0"
tabulate = "^0.9.0"
jsonpath-ng = "^1.6.1"
torch = "2.2.2"
numpy = "1.26.4"
httpx = "0.27.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
deptry = "^0.21.1"
pytest-mock = "^3.14.0"
types-aiofiles = "^24.1.0.20240626"
motor-types = "^1.0.0b4"
pytest-asyncio = "^0.24.0"
requests-mock = "^1.12.1"
mypy = "^1.13.0"
pytest = "^8.3.4"
autoflake = "^2.3.1"
streamlit = "^1.41.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
